docs: add flow chart visualisation
Modify `2021-07-22-Millions-of-UK-residents-struggle-to-access-food`
for better interpretation of the chart.
