+++ content/visualisation/2021-07-22-Millions-of-UK-residents-struggle-to-access-food/index.mdx
@@ -2,7 +2,7 @@
 type: visualisation
 author: [Yu Liang Weng]
 title: Millions of UK residents struggle to access food
-thumbnail: thumb.jpg
+thumbnail: thumb.png
 description: >
   A new study from the University of Sheffield has revealed the areas in the UK where residents most struggle to afford or access food.
   For the first time researchers were able to identify food insecurity at local authority scale across three categories, from those experiencing hunger, to those just one emergency away from going without food.
