+++ content/visualisation/2021-08-04-Which-Statistical-Test-To-Use-For-Two-Variables/flowChart/textData.js
@@ -1,0 +1,26 @@
+export const textForHelp = [
+  {
+    question: 'Introduction',
+    answer: 'This flow chart is a simple tool that help you to choose what statistical test to use for two variables. This was prepared for the Statistical Modelling learning path at dataviz.shef.ac.uk.'
+  },
+  {
+    question: 'How do I get started?',
+    answer: 'To get started, click on the \'Click to start\' animated triangle button. Click on the node (orange colour diamond shape) with question to see more options. Then you can follow options on different lines and click on the next node. Repeat this process until you have reached a rectangle node.'
+  },
+  {
+    question: 'Unselect nodes/shapes',
+    answer: 'Selected nodes will be highlighted in different colours, click it again if you want to unselect the target node.'
+  },
+  {
+    question: 'How do I move around the chart?',
+    answer: 'You can drag the background to move around, and scroll up and down to zoom in or out. Or instead, click on the buttons on the bottom left which provides the same functionality.'
+  },
+  {
+    question: 'Shapes/nodes disappear',
+    answer: 'If nothing shows up on the panel, you can either click on the \'window\' button (just under the zoom out button) on the bottom left, or the \'Fit view\' button on the right panel to restore the view.'
+  },
+  {
+    question: 'License',
+    answer: 'This flow chart and its content are covered by a Creative Commons license CC BY-SA 4.0.'
+  }
+]
